fix compilation & linking errors
